ci: avoid running git diff after patching
Drop `git diff` command so it is easier to run CI locally if git checkout is a
worktree. Currently it fails because the directory is not recognized as a git
repository.

The `git diff` command was added recently in #28359 commit
a21e497ceed461bd828218333c2923dc7a7acedf and can be avoided just by teeing the
patch to stdout