Move BlockMerkleBranch back to merkle.{h,cpp}
The Mining interface uses this function in the next commit
to calculate the coinbase merkle path. Stratum v2 uses
this to send a compact work template.

This partially undoes the change in 4dd8ef3cfe614cdbe2f7477a6d0ccac811b8f22a,
but is not a revert, because the implementation changed in the meantime.

This commit also documents the function.