doc: warn that CheckBlock() underestimates sigops
Counting sigops in the witness and for p2sh requires
context that CheckBlock()  does not have, so it only
counts a subset of sigops.

The check here was introduced by Satoshi as a "cleanup" in
230ab34cb15b35d9da249d49b50b35f8a38d7f85. With the attempted
introduction of OP_EVAL, it was replaced by the check in
ConnectBlock(). Commit 4c5b5d7e8233445be883fc44ddaca0a5c0aafdfd
marked this code as a placeholder for backward compatibility.

Then when P2SH replaced OP_EVAL in 71fbb653be9b27ba3d89f0182ec1ad644c31108b
the phrase "compatibility-breaking" was replaced by a simple
observation that before v0.6 this is how sigops were counted.

It's unclear why the check was kept and there were no review comments
about it.