doc: warn that CheckBlock() underestimates sigops
Counting sigops in the witness and for p2sh requires
context that CheckBlock()  does not have, so it only
counts a subset of sigops.

The check here was introduced by Satoshi as a "cleanup" in
230ab34cb15b35d9da249d49b50b35f8a38d7f85. With the attempted
introduction of OP_EVAL, it was replaced by the check in
ConnectBlock(). Commit 3dae128acb958f766de15212550a9f681cfef9c6
marked this code as a placeholder for backward compatibility.

Then when P2SH replaced OP_EVAL in 0cbce8c5da6cba4bba3dacc2ff9f3b17267c4487
the phrase "compatibility-breaking" was replaced by a simple
observation that before v0.6 this is how sigops were counted.

It's unclear why the check was kept and there were no review comments
about it.