test: Remove intermittent and presumed fixed tsan race suppressions
LoadWallet was added in commit 0c6dea3f0a81291d451aac668ecc5a8009638a31, which
points to a traceback with BerkeleyBatch in it. Now that BDB is removed,
this can be removed as well.

The race in DatabaseBatch was added in commit
b4551aa2d1ada948e1cd5c90f53e36d3f511b2e9, which does not point to a
traceback. This was likely also fixed with the BDB removal.

If not, the suppressions should be added back, mentioning that they are
intermittent and including a traceback and possibly steps to reproduce.