test: Remove intermittent and presumed fixed tsan race suppressions
LoadWallet was added in commit 00e63390b854d60c6521ea8fff97eefa0960e7fe, which
points to a traceback with BerkeleyBatch in it. Now that BDB is removed,
this can be removed as well.

The race in DatabaseBatch was added in commit
ec1e84342f1a752253d39e72126da2c12e26e08c, which does not point to a
traceback. This was likely also fixed with the BDB removal.

If not, the suppressions should be added back, mentioning that they are
intermittent and including a traceback and possibly steps to reproduce.