rpc, wallet: replace remaining hardcoded output types with `FormatAllOutputTypes`
This commit takes use of the `FormatAllOutputTypes` helper
(introduced in PR #32432, commit b7c6f313c8d8871e35d17f4150ea92177cf7b102)
to get rid of the hardcoded output types in wallet RPC documentation.
Note that it can't be used in the `createmultisig` RPC, as this one is
only for pre-taproot output types and hence doesn't contain "bech32m" in the list.